IOX-#21 Move all methods inside the IntrospectionApp and introduce clearToBottom()
