+++ tools/introspection/include/iceoryx_introspection/introspection_app.hpp
@@ -15,8 +15,14 @@
 #pragma once
 
 #include "iceoryx_introspection/introspection_types.hpp"
+#include "iceoryx_posh/popo/subscriber.hpp"
+#include "iceoryx_posh/roudi/introspection_types.hpp"
 
 #include <getopt.h>
+#include <map>
+#include <ncurses.h>
+#include <unistd.h>
+#include <vector>
 
 namespace iox
 {
@@ -39,10 +45,28 @@ static constexpr iox::units::Duration MIN_UPDATE_PERIOD = 500_ms;
 static constexpr iox::units::Duration DEFAULT_UPDATE_PERIOD = 1000_ms;
 static constexpr iox::units::Duration MAX_UPDATE_PERIOD = 10000_ms;
 
+/// @brief color pairs for terminal printing
+enum class ColorPairs : uint8_t
+{
+    redOnBlack = 1,
+    whiteOnRed
+};
+
+/// @brief formatting options for terminal printing
+static const std::map<PrettyOptions, uint32_t> prettyMap = {
+    {PrettyOptions::title, A_BOLD | COLOR_PAIR(static_cast<uint8_t>(ColorPairs::redOnBlack))},
+    {PrettyOptions::highlight, A_BOLD | A_UNDERLINE},
+    {PrettyOptions::error, A_BOLD | COLOR_PAIR(static_cast<uint8_t>(ColorPairs::whiteOnRed))},
+    {PrettyOptions::bold, A_BOLD},
+    {PrettyOptions::normal, A_NORMAL}};
+
+
 /// @brief base class for introspection
 class IntrospectionApp
 {
   public:
+    using SubscriberType = iox::popo::Subscriber;
+
     int updatePeriodMs = DEFAULT_UPDATE_PERIOD.milliSeconds<int>();
     bool doIntrospection = false;
 
@@ -80,6 +104,59 @@ class IntrospectionApp
 
     /// @brief this is needed for the child classes to extend the parseCmdLineArguments function
     IntrospectionApp() noexcept;
+
+    void runIntrospection(const int updatePeriodMs, const IntrospectionSelection introspectionSelection);
+
+  private:
+    /// @brief initializes ncurses terminal
+    void initTerminal();
+
+    /// @brief Erase from courser to the end of the screen
+    void clearToBottom();
+
+    /// @brief closes ncurses terminal
+    void closeTerminal();
+
+    /// @brief prints buffered data to the terminal
+    void refreshTerminal();
+
+    /// @brief updates the first pad coordinates to display
+    void updateDisplayYX();
+
+    /// @brief waits until user input or timeout
+    /// @param[in] timeoutMs timeout in milliseconds (-1 to wait forever)
+    void waitForUserInput(int32_t timeoutMs);
+
+    /// @brief prints to the terminal
+    /// @param[in] str string to print
+    /// @param[in] pr formatting options
+    void prettyPrint(const std::string& str, const PrettyOptions pr = PrettyOptions::normal);
+
+    /// @brief prints active process IDs and names
+    void printProcessIntrospectionData(const ProcessIntrospectionFieldTopic* processIntrospectionField);
+
+    /// @brief prints table showing current mempool usage
+    void printMemPoolInfo(const MemPoolIntrospectionTopic& topic);
+
+    void printPortIntrospectionData(const std::vector<ComposedSenderPortData>& senderPortData,
+                                    const std::vector<ComposedReceiverPortData>& receiverPortData);
+
+    bool waitForSubscription(SubscriberType& port);
+
+    std::vector<ComposedSenderPortData>
+    composeSenderPortData(const PortIntrospectionFieldTopic* portData,
+                          const PortThroughputIntrospectionFieldTopic* throughputData);
+
+    std::vector<ComposedReceiverPortData>
+    composeReceiverPortData(const PortIntrospectionFieldTopic* portData,
+                            const ReceiverPortChangingIntrospectionFieldTopic* receiverPortChangingData);
+
+    /// @brief ncurses pad
+    WINDOW* pad;
+    /// @brief first pad row to show on the ncurses window
+    int32_t yPad{0};
+    /// @brief first pad column to show on the ncurses window
+    int32_t xPad{0};
 };
 
 } // namespace introspection
